azure_ocr_validador_bono/requirements.txt
flask
python-dotenv
requests
pymupdf
pyjwt
